Add a registry reporter
Add a new cli tool that can cycle through a the catalog of a
registry (with filtering ability) and submit a debmonitor report for
all of the latest image tags found.

It can get filter rules from the command line or from a configuration
file where more complex rules are allowed.

Change-Id: Idefedc07fdeee28eea7193f4bfd854c11ed10dc3