Fix tox / pep517 issues
When using pip > 1.19, tox will fail if a pyproject.toml file is present
and usedevelop is set to true and we don't fully migrate to it from
setup.py.

Given I am not interested in the migration yet, I prefer to rename the
black configuration file and to pass it explicitly to tox.

Change-Id: I7f7c8266d8c6f032d002701eda3d67a5d0eab53a