Add a registryctl command-line utility
It is handy sometimes to be able to query and operate programmatically on a
docker registry.
The docker client doesn't allow to do that with ease, so registryctl is a
good replacement for our own usage.

For now, registryctl only has the following abilities:

- list all images and their tags on a registry, in yaml format
- list all tags for an image, filtered via globbing
- delete specific versions of an image, again filtered via globbing

In order to do this, registry_browser was refactored to a more general registry
class and a couple subclasses for different functionalities.

Also:

* added full branch coverage reporting
* enabled typing checks

this means there are additional changes all across the repository.

Bug: T242604
Change-Id: I15ae9a3759d3696fcce42ebf84f5ee023c0d8a57