Allow to only report images of supported Debian versions
Excluding unsupported Debian versions is currently done by complex
filter rules based on images names etc. That is error prone and
dangerous as image names stay the same while tags (and with them the
base image versions) change.

This change implements parsing of the debian version an image is based
on and adds an --minimum-debian-version option (defaults to 10) to
ignore all images below this version.
The downside of this is that the images sill need to get pulled for the
version to be checked. But it should prevent us from to broad excludes
by filter rules.

Bug: T348876
Bug: T335282
Change-Id: I66f3376ce04f4099615c9f7e82338c9f74def25a