Add name_filters support to the k8s browser
At the moment we don't have any name_filter support for the k8s
browser, and we are starting to have some use cases where it
would be beneficial.

Add a basic support that tries to match the name of the docker image
not considering the URL/site bits and the tags (as we do for
the registry browser).

Change-Id: I0eb6c2c133e1f2a893067e78315426a8f3d6b730